[ENH] move utility functions to utils.py (#356)
