+++ janitor/functions.py
@@ -13,42 +13,16 @@ from scipy.stats import mode
 from sklearn.preprocessing import LabelEncoder
 from typing import Callable, Dict, Iterable, List, Union, Any, Tuple
 
+from .utils import (
+    deprecated_alias,
+    _strip_underscores,
+    check,
+    _clean_accounting_column,
+    _currency_column_to_numeric,
+    _replace_empty_string_with_none,
+    _replace_original_empty_string_with_none,
+)
 from .errors import JanitorError
-from .utils import deprecated_alias
-
-
-def _strip_underscores(
-    df: pd.DataFrame, strip_underscores: Union[str, bool] = None
-) -> pd.DataFrame:
-    """
-    Strip underscores from DataFrames column names.
-
-    Underscores can be stripped from the beginning, end or both.
-
-    .. code-block:: python
-
-        df = _strip_underscores(df, strip_underscores='left')
-
-    :param df: The pandas DataFrame object.
-    :param strip_underscores: (optional) Removes the outer underscores from all
-        column names. Default None keeps outer underscores. Values can be
-        either 'left', 'right' or 'both' or the respective shorthand 'l', 'r'
-        and True.
-    :returns: A pandas DataFrame with underscores removed.
-    """
-    underscore_options = [None, "left", "right", "both", "l", "r", True]
-    if strip_underscores not in underscore_options:
-        raise JanitorError(
-            f"strip_underscores must be one of: {underscore_options}"
-        )
-
-    if strip_underscores in ["left", "l"]:
-        df = df.rename(columns=lambda x: x.lstrip("_"))
-    elif strip_underscores in ["right", "r"]:
-        df = df.rename(columns=lambda x: x.rstrip("_"))
-    elif strip_underscores == "both" or strip_underscores is True:
-        df = df.rename(columns=lambda x: x.strip("_"))
-    return df
 
 
 @pf.register_dataframe_method
@@ -2236,110 +2210,6 @@ def reset_index_inplace(df: pd.DataFrame, *args, **kwargs) -> pd.DataFrame:
 
     df.reset_index(*args, **kwargs)
     return df
-
-
-def check(varname: str, value, expected_types: list):
-    """
-    One-liner syntactic sugar for checking types.
-
-    Should be used like this::
-
-        check('x', x, [int, float])
-
-    :param varname: The name of the variable.
-    :param value: The value of the varname.
-    :param expected_types: The types we expect the item to be.
-    :returns: TypeError if data is not the expected type.
-    """
-    is_expected_type = False
-    for t in expected_types:
-        if isinstance(value, t):
-            is_expected_type = True
-            break
-
-    if not is_expected_type:
-        raise TypeError(
-            "{varname} should be one of {expected_types}".format(
-                varname=varname, expected_types=expected_types
-            )
-        )
-
-
-def _clean_accounting_column(x: str) -> float:
-    """
-    Perform the logic for the `cleaning_style == "accounting"` attribute.
-
-    This is a private function, not intended to be used outside of
-    ``currency_column_to_numeric``.
-
-    It is intended to be used in a pandas `apply` method.
-
-    :returns: An object with a cleaned column.
-    """
-    y = x.strip()
-    y = y.replace(",", "")
-    y = y.replace(")", "")
-    y = y.replace("(", "-")
-    if y == "-":
-        return 0.00
-    return float(y)
-
-
-def _currency_column_to_numeric(x, cast_non_numeric=None) -> str:
-    """
-    Perform logic for changing cell values.
-
-    This is a private function intended to be used only in
-    ``currency_column_to_numeric``.
-
-    It is intended to be used in a pandas `apply` method, after being passed
-    through `partial`.
-    """
-    acceptable_currency_characters = {
-        "-",
-        ".",
-        "1",
-        "2",
-        "3",
-        "4",
-        "5",
-        "6",
-        "7",
-        "8",
-        "9",
-        "0",
-    }
-    if len(x) == 0:
-        return "ORIGINAL_NA"
-
-    if cast_non_numeric:
-        if x in cast_non_numeric.keys():
-            check(
-                "{%r: %r}" % (x, str(cast_non_numeric[x])),
-                cast_non_numeric[x],
-                [int, float],
-            )
-            return cast_non_numeric[x]
-        else:
-            return "".join(i for i in x if i in acceptable_currency_characters)
-    else:
-        return "".join(i for i in x if i in acceptable_currency_characters)
-
-
-def _replace_empty_string_with_none(x):
-    if isinstance(x, int):
-        return x
-
-    elif isinstance(x, float):
-        return x
-
-    elif len(x):
-        return x
-
-
-def _replace_original_empty_string_with_none(x):
-    if x != "ORIGINAL_NA":
-        return x
 
 
 @pf.register_dataframe_method
